surface/requirements.txt
# Core Libraries
Django==5.2.4
django-admin-rangefilter==0.11.0
django-after-response==0.2.2
django-object-actions==4.2.0
djangorestframework==3.14.0
django-restful-admin==1.1.3
djangorestframework-queryfields==1.0.0
django-filter==2.4.0
django-import-export==4.3.9
django-nested-admin==4.0.2
django-daterangefilter==1.0.0
django-jsoneditor==0.2.4
netaddr==0.8.0
django-unfold==0.68.0

# our own
django-dbcleanup==0.1.5
django-logbasecommand==0.0.5
django-notification-sender==0.1.0
django-dkron==1.2.2
django-slack-processor==0.1.9
django-olympus==0.0.6
django-environ==0.10.0
django-impersonator==0.0.3
django-apitokens==0.0.3
django-sbomrepo==0.0.10

mysqlclient==2.2.4
tqdm==4.65.0  # for core_utils that is not really a app/package ..?
django-database-locks==0.5  # distributed locks (on mysql)
django-bulk-update-or-create==0.3.0  # for faster batch operations with update_or_create
django-dynamicsettings==0.0.3

-r scanners/requirements.txt
-r sca/requirements.txt
